refactor: extract config path helpers
