+++ src/config.rs
@@ -7,6 +7,9 @@ use std::path::PathBuf;
 use chrono::{Local, NaiveDate};
 use serde::{Deserialize, Deserializer, Serialize};
 
+mod paths;
+use paths::*;
+
 /// Persistent application configuration stored as TOML.
 ///
 /// File locations:
@@ -846,39 +849,6 @@ pub(crate) fn app_data_path(file_name: &str) -> Option<PathBuf> {
 
 fn app_dir() -> Option<PathBuf> {
     app_dir_with_env(|key| std::env::var_os(key))
-}
-
-fn app_dir_with_env(get_var: impl FnMut(&str) -> Option<OsString>) -> Option<PathBuf> {
-    let config_dir = config_dir_from_env(get_var)?;
-    Some(config_dir.join("focustime"))
-}
-
-fn config_dir_from_env(mut get_var: impl FnMut(&str) -> Option<OsString>) -> Option<PathBuf> {
-    #[cfg(target_os = "windows")]
-    {
-        env_path_from_value(get_var("APPDATA")?)
-    }
-    #[cfg(not(target_os = "windows"))]
-    {
-        // Honour XDG_CONFIG_HOME if set, otherwise fall back to ~/.config.
-        if let Some(xdg) = get_var("XDG_CONFIG_HOME").and_then(env_path_from_value) {
-            return Some(xdg);
-        }
-        let home = get_var("HOME").and_then(env_path_from_value)?;
-        Some(home.join(".config"))
-    }
-}
-
-fn env_path_from_value(value: OsString) -> Option<PathBuf> {
-    if value.is_empty() {
-        return None;
-    }
-    if let Some(value_utf8) = value.to_str()
-        && value_utf8.trim().is_empty()
-    {
-        return None;
-    }
-    Some(PathBuf::from(value))
 }
 
 fn nonzero_or_default_u64(value: u64, default: u64) -> u64 {
